cast to size_t

+++ arangod/IResearch/IResearchFeature.cpp
@@ -1152,7 +1152,7 @@ void IResearchFeature::reportRecoveryProgress(arangodb::IndexId id,
       now - _progressState.lastReportTime >= std::chrono::minutes(1)) {
     // report progress only when link id changes or one minute has passed
 
-    auto progress = static_cast<int>(100.0 * current / total);
+    auto progress = static_cast<size_t>(100.0 * current / total);
     LOG_TOPIC("d1f18", INFO, TOPIC)
         << "recovering arangosearch link " << id << ", " << phase
         << ": operation " << (current + 1) << "/" << total << " (" << progress
